Work around a consistent timeout issue in Hinrik's POE::Wheel::Run test.
There is a mysterious 5sec delay when starting processes on some
Windows machines.  The consistent, roundish number of seconds makes me
suspect something environmental, like an antivirus program, but that's
just a wild guess.

Thanks to ultramantis on irc.perl.org #poe or letting me access their
machine with teamviewer.  I couldn't reproduce the problem on my own
systems, so it was vital to see it failing somewhere.