Very large select() timeouts can trigger EINVAL on Mac OSX.
Sérgio Bernardino reported that POE::Component::Server::IRC's synopsis
would go into a busy loop when a client tried to connect.  It turns
out that POE::Loop::Select would fail when the component set a timeout
for a very long time in the future.  Cap select() timeouts to an hour
to avoid the problem.