Move the signal information out of the signal pipe.
The signal pipe included all signal information.  This reduces the
number of signals that can be dispatched in a single event loop
iteration.  It also limits the amount and/or type of information we
can include in the signal.

This change puts the signal information in a plain Perl list and only
uses the signal pipe to wake up the event loop to dispatch the signals
in a safer place.